wordle_bot/requirements.txt
click==8.1.2
MouseInfo==0.1.3
numpy==1.22.3
packaging==21.3
Pillow==9.1.0
pkg_resources==0.0.0
PyAutoGUI==0.9.53
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyparsing==3.0.7
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.28
pytesseract==0.3.9
python3-xlib==0.15
pytweening==1.0.4
